DataViews: Adds a control to the views actions to switch layouts
