+++ packages/edit-site/src/components/dataviews/index.js
@@ -1,2 +1,1 @@
 export { default as DataViews } from './dataviews';
-export { PAGE_SIZE_VALUES } from './view-actions';
