+++ packages/edit-site/src/components/dataviews/text-filter.js
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
@@ -15,18 +10,19 @@ import { SearchControl } from '@wordpress/components';
  */
 import useDebouncedInput from '../../utils/use-debounced-input';
 
-export default function TextFilter( {
-	className,
-	searchLabel = __( 'Filter list' ),
-	onChange,
-} ) {
-	const [ search, setSearch, debouncedSearch ] = useDebouncedInput();
+export default function TextFilter( { view, onChangeView } ) {
+	const [ search, setSearch, debouncedSearch ] = useDebouncedInput(
+		view.search
+	);
 	useEffect( () => {
-		onChange( debouncedSearch );
-	}, [ debouncedSearch, onChange ] );
+		onChangeView( ( currentView ) => ( {
+			...currentView,
+			search: debouncedSearch,
+		} ) );
+	}, [ debouncedSearch, onChangeView ] );
+	const searchLabel = __( 'Filter list' );
 	return (
 		<SearchControl
-			className={ classnames( 'dataviews__text-filter', className ) }
 			onChange={ setSearch }
 			value={ search }
 			label={ searchLabel }
