+++ packages/edit-site/src/components/dataviews/view-grid.js
@@ -1,0 +1,3 @@
+export function ViewGrid() {
+	return 'Grid';
+}
